requirements.txt
fpdf2
selenium